refactor: move nullable into internal
